requirements.txt
# Build tools
setuptools>=65.0.0
wheel>=0.38.0

# Core dependencies
streamlit>=1.29.0
python-dotenv>=1.0.0

# RAG components
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2

# Document processing
pypdf>=3.17.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Embeddings and LLM
groq>=0.4.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Utilities
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
